requirements.txt
fastai==2.7.16
geopandas==1.0.1
imutils==0.5.4
ipython==8.12.3
matplotlib==3.9.2
numpy==2.1.0
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
pandas==2.2.2
Pillow==10.4.0
rasterio==1.3.10
scikit_learn==1.5.1
seaborn==0.13.2
Shapely==2.0.6
skimage==0.0
streamlit==1.37.1
supervision==0.23.0
torch==2.4.0
tqdm==4.66.5
ultralytics==8.2.57
